Simplifying memory globalization from the front end to move optimizations to the middle end.
Memory globalization was fully implemented in the front end. There are three runtimefunctions in Libomptarget:

* __kmpc_data_sharing_push_stack
* __kmpc_data_sharing_coalesced_push_stack
* __kmpc_data_sharing_pop_stack

The front end performed a scape analysis and created a record declare with all the stackvariables.
Then, based on the context (isTTD and other parameters) it would create a pushfor the size of the record,
or for that size multiplied by the WARP (to globalize for thewhole WARP.This PR removes the record creation,
and it simplifies the front end to be a simple runtimecall that will be later on optimized in the middle end.
The middle end will be able todetermine the stack variables that do scape, and those that do not, as well as
theapprorpiate merging of different globalized variables