Frist driver implementation:
This commit adds the target triple "colossus-graphcore-unknown-elf" necessary for
the graphcore architecture. It also includes a device toolkit that calls the popc
compiler on the .ll file. It also have a copy of the NVPTX Code Gen file as a placeholder
for the front end that will generate the codelet graph.

This can be used like this:
clang a.c -fopenmp -fopenmp-targets=colossus-graphcore

It will currently fail because popc does not recognize the output of CodeGen completely

For more debug info:
clang -ccc-print-bindings a.c -fopenmp -fopenmp-targets=colossus
clang -ccc-print-phases a.c -fopenmp -fopenmp-targets=colossus
clang -### a.c -fopenmp -fopenmp-targets=colossus