Adding first runtime implementation
Adding first runtime implementation taken from the files that already exist in UpDown.
Adding also support for multiple functions, each function will be added to the
final function in python. The final function in python uses the
file name.
The order in which events are added is the same order it shows in the file. Therefore variables are shared across ffunctions