Adding Runtime test, Fixing CMakeLists for tests, changing memory map
Commit adds a new runtime test that dumps the address base
pointers and MAchineConfiguration parameters

This also fixes some previously introduced CMakeLists errors due to
the use of clang-format on the CMakeLists.txt file

Finally, the memory address space has completely changed to adapt to
the accorded addresses. This separates control and SPMem completely
and makes the start values parametric. Based on the new address
the runtime has been changed to use the new parametric values