Adding simruntime, minor changes and fixes
This branch adds the simruntime class. This allows to create a wrapper around the
UDruntime_t class that allocates memory to allow to test the interfaces.
In the future this simruntime class will call the emulator directly, to allow for
functional emulation of UpDown programs.

Additionally, this commit changes the configuration parameters
for the offsets in the control registers, from bytes, to words. This
allows pointer arithmetic to be easier.

Finally, minor fixes in the documentations and the CMakeLists.txt files