First cleaned up version of CodeGen for Colossus.
The original code gen was an exact copy of GPU code gen. I have removed most of the
GPU specific fucntions and unnecessary code gen for reductions and others.
This implies that support for many openMP features within the target region is now
Not allowed. This is not heavily tested.