Adding Jinja2CppLight files
I have added the files from the Jinja2CppLight repository. I have
decided not to use a git submodule since I have already made changes to
the files to move some of the definitions to the source file and not the
header file. Therefore, there is already a broken version that will
require additional efforts to support. This is just for the sake of the
syntax plugin, I should remove this dependency in stand alone
installation