Moving a function to the runtime API
Changed get_lane_local_memory to get_lane_physical_memory. As this is used by
the ubenchmarks, this function is required right now since no mechanism to mapping
virtual to physical memory exists. There is no malloc for the scratchpad memory either
therefore, this function returns a pointer that corresponds to the phyisical memory location
in the scratchpad memory.

This funciton was moved to UDRuntime_t
Removed the change endianess function.