Fixing many bugs, adding tests, minor improvements
- Implementing reset_mapped_memory() to be used during simulation to reset the regions
of the mapped memory to be mapped to the allocated memory address space.
- Fixing a problem during operands_t creation that required a copy with sizeof(word_t)
- Fixing order in memcpy for mm2t_memcpy and t2mm_memcpy.
- During event simulation, recursive was changed to be delayed, as it is not possible to start a new
event in the same lane. The lanes to be processed are kept in a vector of pairs
that is later on used before returning from the execute function. This is because the memory
operations are atomic, and they create an event right away.
- Changing the implementation of memory mapped access from updown since mm2t and t2mm
actually map to an offset within the region, but what we have is pointers. So we need to
access these pointers directy. This is for the emulator
- Added tests for memcpy both only from the top code as well as using the emulator
Working code demonstrates the updown peforms the copy operation.