Implementing trace mode for the emulator
The emulator implements the EFAUtil print debug level. This was controlled
by changing a line in python. Say no more! this value is set directly from C++
Through the interface. The constructor for the Simruntime will receive
A flag in the form of an enum that indicates the level of debugging

This also cleans up two functions that were not needed in the emulator
c++ interface: those related to sbuffer. This needs to be fixed in the emulator