Changing _memcpy to use a void* instead of a word_t pointer
I have decided that it makes more sense to always try to work
With bytes, instead of words. memcpy already used bytes for the offset,
adding the word in the memcpy, made it more difficult to understand
This changes all the signatures of the memcpy APIs to use void* and
let the user determine the size of data to be copied.

This patch also updates all the tests that use this interface. I have done
Quick result checks, but it would be a good idea to look out for other errors that
may show up