Changing ANY_THREAD for CREATE_THREAD
ANY_THREAD is confusing and does not accurately describes
that this event creation will be assigned to a new context thread.
To fix this I have replace all the ANY_THREAD references to
CREATE_THREAD before it is too late.

I also remove a variable that is not needed anylonger.