Bug fix: Events not sent after yeld_terminate
The execution of a yield_terminate was bypassing the events sent during
that particular activation of the thread. This is because the number of events
sent was being taken from the stats, and stats were being ignored in the
original implementation. This patch fixes this issue, by renaming the stats
and using its value to determine if there are events to be sent. This makes
the mapreduce program execute without any issues.