Implementing non-blocking scheduling
The previous implementation will stuck the top thread
until no more events were executable. However, this has a
drawback. When the lanes are expecting for a code to exec
in the top, it may never happen. We return after 100 scheduling
iterations, and allow the top to continue. We re-schedule if the
top calls test_ or test_wait calls. While it is still possible
to write a program that does not finish in the updown,
it is often the case that termination is check with a
spin wait loop. We allow this to be an admissable limitation
of the emulator.

This also includes a test that would previously block the emulator
but right now it does not. It also fixes some minor errors
in the interaction with the emulator when the number of
operands in the event is 0.