Deleting temp files at the end of execution
In order to communicate C++ to Python, we use mmap on
some files that connect the two worlds. These files were being
Created but never removed. The content of this file is not useful after
execution has finished. We add logic to remove this files during
Destructor of the runtime

Adding appropriate runtime delete to all the tests.