Adding env variable to set scheduling iterations
We have added the environment variable `UPDOWN_SIM_ITERATIONS`
to specify how many scheduling iterations are performed
when the control is passed to the UpDown. The default
value is 100. if `UPDOWN_SIM_ITERATIONS` is set to 0, then the
control will only return from the UpDown whenever there are
no more events to schedule. This was the original scheduling policy.