Merged in pagerank (pull request #2)
Pagerank

* Placeholder for apps, including TC and Map Reduce

Adding the apps folder, and its corresponding CMakeFile logic
to build. This uses the UPDOWNRT_ENABLE_APPS Cmake flag to
enable/disable compilation of apps.

* Adding libraries for graph computation

Adding:
SNAP and glib

* Adding the placeholder for the pagerank app

This removes the lines in CMakeFile that refer to triangle
count, but we are using the branch forking from TC because
it already has the snap folders. During merge we will likely have to
make some slight changes…
* pagerank finished, simupdown.h modified adding string lib

* Adding the Snap.o to the compilation of the apps

It is necessary to compile the program. The .o is used to keep
the linking static, such that we do not rely on dynamic libraries
when using the Gem5 simulator

* pagerank with cpp runtime emu code complete.

the code now can run successfully,
see code comments for todos

* Merged master into pagerank
* Pagerank code cleanup.
Added detailed comment.

* pagerank unused and bin file removed.