Implementing multi UD emulation
This commit implements multi UD by changing the event word to contain,
updating the interface to create an emulator per UD, and updating the
python emulator implementation to fix necessary changes. Furthermore,
it introduces a new test that is a multiUD code for memcpy. Below summary
of all the changes.

- The SimUpDown now creates a python interface per UD, resulting in array of
interfaces. Each interface is independently managed, but its functional
behavior is all the same
- The Eventword now contains an structure that encodes the UD and LaneID.
Instead of using 8 bits for the laneID, the LaneID is now encoded in 6 bits and
the updown is encoded in the remaining 2 bits in the MSB of the old Lane ID.
For example 0b11000001 is lane 1 of UD 3. This means that this version can
only emulate up to 4 up downs.
- The name of the file used for mmap also contains the UD ID. Instead of
lane##_send.txt it is now XXlane##_send.txt where XX is the UDID
- The interface has been updated to consider multiUD. Furthermore, we have
added a wrapper for error handling of python code for an improved error
message display whenever there is an issue in Python.
- The EFA, EfaExecutor_multilane_nothread and EfaUtil have been updated
to use a version that is closer to that one use in Gem5. This to consider
recent changes to improve performance, as well as to be multiUD ready
- A new memcpyEFA.py is implemented that allows for multiUD and multilane
memcpy. A new test that uses this file is added that spawns 4 UDs with 4 lanes
each, and peforms a multiUD and MultiLane simulation. The name of this test is
`simrt_pyEmu_mm_memcpy_multiud.cpp`.