placeholder implementation for CGOpenMPRuntimeColossus
This is the first clean up and placeholder implementation
for the Code Gen OpenMP Runtime for the Colossus
target. This implementation is divided in two parts:

- A ColossusCodeGenFunction class. A wrapper class for CodeGenFunction
that allows to change the behavior for several clang::Stmt. This is necessary
because basic Stmt such as if, while, etc. within the target region must be
lowered into poplar API function calls. There are also limitations on what can
go into the API.
This also changes the behavior of the OMPTaskDirectiveClass. The complexity is that
the target region itself should not emit anything. We need to emit the codelets
(i.e., only the tasks within the target region) as functions. Therefore, we remove the
outlined function, together with the task entry point insertion. We jump
Straight into codegen for that body. Later on this will be further tunned to change
the variable types and connect to the rest of the API

- CGOpenMPRuntimeColossus: This is the class for code generation of the OpenMP
runtime for the colossus architecture. It is used when generating code for the
colossus triple.

A lot of the code is simplified placeholders. There are still print messages, but we
can start dividing up work here.