Placeholder to allow the CGopenMPRuntimeColossus to generate host code
This allows to use the right codegen runtime for making device specific
changes in the code generation for OpenMP. As is, this means that the host code
will not generate anything other than the outlined regions. These will later on
be ignored since we do not yet plan to support host fall back mechanism.