Placeholder for DeviceRTLty
For the device RTL Type, we have to start creating the data structures
that hold the poplar program, target, etc.

This is the first step towards that
Two data structures are created DeviceRTLty and Program. The poplar
progam will serve as a wrapper for the definition of a poplar graph
that has an associated engine and program to it. grupping all these together.

Next step is to start creating the right runtime APIs.