Integrating the Colosus Backend.
This patch takes all the changes from the llvm-project-fork
from graphcore that are related to the colossus backend,
and integrates them into this version of the compiler.

This is to allow for the codegen in OpenMP to generate the
right attributes and cpu types. Given my lack of knowledge of the
backend I rather include most of it and make sure it builds. This ignores
completely all the tests (not so smart but in the interest of time)

Original repo:
https://github.com/graphcore/llvm-project-fork

ALL CHANGES IN THIS COMMIT ARE FROM GRAPHCORE